chore(sdk): generate csharp server sdk (230524.2314)
- generated from source sdk spec commit: 0a0d134d7938a8c701b1c7181ba19b04df0b2a81

+++ apis/AccelByte.Sdk.Api.Dsmc/Model/ModelsCreateImageRequest.cs
@@ -22,7 +22,7 @@ namespace AccelByte.Sdk.Api.Dsmc.Model
         public string? Image { get; set; }
 
         [JsonPropertyName("imageSize")]
-        public int? ImageSize { get; set; }
+        public long? ImageSize { get; set; }
 
         [JsonPropertyName("namespace")]
         public string? Namespace { get; set; }
